edit endpoint_mapping to hold all body mapping and endpoint management
