+++ mlrun/runtimes/nuclio/serving.py
@@ -15,11 +15,8 @@ import json
 import os
 from base64 import b64decode
 from copy import deepcopy
-from http import HTTPMethod
 from typing import Union
 
-import jsonpath_ng
-import jsonpath_ng.exceptions
 import nuclio
 from nuclio import KafkaTrigger
 from nuclio.triggers import NuclioTrigger
@@ -31,11 +28,11 @@ import mlrun.datastore.datastore_profile as ds_profile
 import mlrun.runtimes.kubejob as kubejob_runtime
 import mlrun.runtimes.nuclio.function as nuclio_function
 import mlrun.runtimes.pod as pod_runtime
-import mlrun.serving.utils as serving_utils
 from mlrun.datastore import get_kafka_brokers_from_dict, parse_kafka_url
 from mlrun.model import ObjectList
 from mlrun.runtimes.function_reference import FunctionReference
 from mlrun.secrets import SecretsStore
+from mlrun.serving.endpoint_mapping import APIHandlerConfig, BodyMappings, EndpointConfig
 from mlrun.serving.server import (
     GraphServer,
     add_system_steps_to_graph,
@@ -54,384 +51,6 @@ from mlrun.utils import get_caller_globals, logger, merge_requirements, set_path
 
 serving_subkind = "serving_v2"
 
-
-class BodyMappings(mlrun.model.ModelObj):
-    """Directional parameter mappings for a single phase — input (REST → graph) or output (graph → REST).
-
-    The direction is determined by which parameter of ``add_endpoint_handler`` the instance
-    is passed to (``body_mappings`` for input, ``output_body_mappings`` for output), not by
-    the class itself.
-
-    Usage::
-
-        # Input: extract fields from the incoming REST request body
-        input_bm = BodyMappings()
-        input_bm.add_mapping("$.model", destination_path="model", mandatory=True)
-        input_bm.add_mapping("$.messages", destination_path="messages", mandatory=True)
-        input_bm.add_mapping("$.temperature", destination_path="temperature")
-
-        # Output: reshape the graph response before returning to the caller
-        output_bm = BodyMappings()
-        output_bm.add_mapping(
-            "message.content", destination_path="content", mandatory=True
-        )
-        output_bm.add_mapping("finish_reason", destination_path="finish_reason")
-
-        config = APIHandlerConfig()
-        config.add_endpoint_handler(
-            "/v1/chat/completions",
-            HTTPMethod.POST,
-            body_mappings=input_bm,
-            output_body_mappings=output_bm,
-        )
-    """
-
-    _dict_fields = ["mappings"]
-
-    def __init__(self) -> None:
-        # _by_src is the primary store: src → {"destination_path", "mandatory"}.
-        # _by_dest_index is a reverse lookup: dest → src, for O(1) duplicate detection.
-        # _by_dest_index is empty for validate-only mappings (destination_path=None).
-        self._by_src: dict[str, dict] = {}  # src → {"destination_path", "mandatory"}
-        self._by_dest_index: dict[str, str] = {}  # dest → src (reverse index only)
-
-    @property
-    def mappings(self) -> list[dict]:
-        return [{"source_json_path": src, **data} for src, data in self._by_src.items()]
-
-    @mappings.setter
-    def mappings(self, value: list[dict]) -> None:
-        self._by_src = {}
-        self._by_dest_index = {}
-        for m in value or []:
-            src = m["source_json_path"]
-            dest = m.get("destination_path")
-            self._by_src[src] = {
-                "destination_path": dest,
-                "mandatory": m.get("mandatory", False),
-            }
-            if dest is not None:
-                self._by_dest_index[dest] = src
-
-    def add_mapping(
-        self,
-        source_json_path: str,
-        destination_path: str,
-        mandatory: bool = False,
-    ) -> None:
-        """Add a single field mapping.
-
-        :param source_json_path: JSONPath expression to extract the value from the source.
-                                 For input — extracts from the REST request body (e.g. ``"$.model"``).
-                                 For output — extracts from the graph response (e.g. ``"message.content"``).
-        :param destination_path: Where to place the extracted value.
-                                 For input — the parameter name passed into the graph.
-                                 For output — the field name in the REST response returned to the caller.
-        :param mandatory: If ``True``, a missing field raises an error at request time.
-                          If ``False``, a missing field is included as ``None`` in the output
-                          (output) or silently skipped (input).
-        :raises mlrun.errors.MLRunInvalidArgumentError: If ``destination_path`` is empty.
-        """
-        if not source_json_path:
-            raise mlrun.errors.MLRunInvalidArgumentError(
-                "source_json_path must be a non-empty string"
-            )
-        if not destination_path:
-            raise mlrun.errors.MLRunInvalidArgumentError(
-                "destination_path must be a non-empty string"
-            )
-        try:
-            jsonpath_ng.parse(source_json_path)
-        except (
-            jsonpath_ng.exceptions.JsonPathLexerError,
-            jsonpath_ng.exceptions.JsonPathParserError,
-        ) as exc:
-            raise mlrun.errors.MLRunValueError(
-                f"Invalid JSON path expression for parameter '{destination_path}': "
-                f"'{source_json_path}'. Error: {exc}"
-            ) from exc
-
-        entry = {"destination_path": destination_path, "mandatory": mandatory}
-
-        # Duplicate source — overwrite existing entry.
-        if source_json_path in self._by_src:
-            old_dest = self._by_src[source_json_path].get("destination_path")
-            mlrun.utils.logger.warning(
-                "Overriding existing body mapping",
-                source_json_path=source_json_path,
-                old_destination=old_dest,
-                new_destination=destination_path,
-            )
-            self._by_dest_index.pop(old_dest, None)
-            self._by_src[source_json_path] = entry
-            self._by_dest_index[destination_path] = source_json_path
-            return
-
-        # Duplicate destination — overwrite existing entry.
-        if destination_path in self._by_dest_index:
-            old_src = self._by_dest_index[destination_path]
-            mlrun.utils.logger.warning(
-                "Overriding existing body mapping",
-                source_json_path=source_json_path,
-                old_destination=destination_path,
-                new_destination=destination_path,
-            )
-            self._by_src.pop(old_src, None)
-            self._by_src[source_json_path] = entry
-            self._by_dest_index[destination_path] = source_json_path
-            return
-
-        self._by_src[source_json_path] = entry
-        self._by_dest_index[destination_path] = source_json_path
-
-    def remove_mapping(self, destination_path: str) -> None:
-        """Remove the mapping with the given destination_path. No-op if not found.
-
-        :param destination_path: The destination key to remove.
-        """
-        if destination_path not in self._by_dest_index:
-            return
-        src = self._by_dest_index.pop(destination_path)
-        self._by_src.pop(src, None)
-
-    def __repr__(self) -> str:
-        return f"BodyMappings(mappings={self.mappings!r})"
-
-
-class EndpointConfig(mlrun.model.ModelObj):
-    """Configuration for a single API endpoint — routing and input mapping in one object."""
-
-    _dict_fields = [
-        "path",
-        "http_method",
-        "action",
-        "description",
-        "input_body_mappings",
-        "output_body_mappings",
-    ]
-
-    def __init__(
-        self,
-        path: str = "",
-        http_method: HTTPMethod | str = HTTPMethod.POST,
-        action: schemas.APIHandlerAction = schemas.APIHandlerAction.ALLOW,
-        description: str | None = None,
-        input_body_mappings: BodyMappings | None = None,
-        output_body_mappings: BodyMappings | None = None,
-    ) -> None:
-        self.path = self._normalize_path(path)
-        self._validate_path(self.path)
-        self.http_method = APIHandlerConfig._validate_http_method(http_method)
-        self.action = action
-        self.description = description
-        self.input_body_mappings = input_body_mappings
-        self.output_body_mappings = output_body_mappings
-
-    @staticmethod
-    def _normalize_path(path: str) -> str:
-        """Normalize path to ensure it starts with a forward slash."""
-        if not path.startswith("/"):
-            return f"/{path}"
-        return path
-
-    @staticmethod
-    def _validate_path(path: str) -> None:
-        """Validate an endpoint path for structural correctness.
-
-        Currently enforces wildcard ``*`` rules:
-
-        * ``*`` may only appear once.
-        * ``*`` must be the last character in the path.
-
-        :param path: Normalized path (with leading ``/``) to validate.
-        :raises mlrun.errors.MLRunValueError: If the path contains an invalid ``*`` pattern.
-        """
-        star_count = path.count("*")
-        if star_count == 0:
-            return
-        if path[-1] != "*":
-            raise mlrun.errors.MLRunValueError(
-                f"Invalid endpoint path '{path}': "
-                f"wildcard '*' must be at the end of the path"
-            )
-        if star_count > 1:
-            raise mlrun.errors.MLRunValueError(
-                f"Invalid endpoint path '{path}': "
-                f"wildcard '*' must appear only once at the end of the path"
-            )
-
-    def get_endpoint_key(self) -> str:
-        """Return the endpoint key in the format 'METHOD:path', e.g. 'POST:/v1/chat/completions'."""
-        return serving_utils.combine_serving_endpoint_key(self.http_method, self.path)
-
-    def __repr__(self) -> str:
-        return (
-            f"EndpointConfig(path={self.path!r}, http_method={self.http_method!r}, "
-            f"action={self.action!r}, input_body_mappings={self.input_body_mappings!r}, "
-            f"output_body_mappings={self.output_body_mappings!r})"
-        )
-
-
-class APIHandlerConfig(mlrun.model.ModelObj):
-    """Configuration for API handler in serving graph"""
-
-    _dict_fields = ["enabled", "endpoints", "include_url_info"]
-
-    def __init__(
-        self,
-        enabled: bool = True,
-        endpoints: dict[str, dict | EndpointConfig] | None = None,
-        include_url_info: bool = False,
-    ):
-        self.enabled = enabled
-        self._endpoints: dict[str, EndpointConfig] = {}
-        self.endpoints = endpoints or {}
-        self.include_url_info = include_url_info
-
-    @property
-    def endpoints(self) -> dict[str, "EndpointConfig"]:
-        """Get the endpoints as a dict keyed by endpoint key (``"METHOD:path"``)."""
-        return self._endpoints
-
-    @endpoints.setter
-    def endpoints(self, endpoints: dict[str, dict | EndpointConfig]) -> None:
-        """Set the endpoints from a dict of raw dicts (deserialization) or EndpointConfig objects."""
-        self._endpoints = {}
-        for endpoint_key, ep in endpoints.items():
-            if isinstance(ep, EndpointConfig):
-                self._endpoints[endpoint_key] = ep
-            else:
-                # dict → deserialization path. Key is "METHOD:path" (e.g. "POST:/v1/chat/completions"),
-                # a derived cache for O(1) lookup — authoritative data lives in EndpointConfig itself.
-                # Old format: {"action": "allow", "description": "..."}
-                # New format adds: "path", "http_method", "input_body_mappings"
-                # Manually deserialize nested BodyMappings since ModelObj does not do it automatically
-                if "path" not in ep or "http_method" not in ep:
-                    raise mlrun.errors.MLRunInvalidArgumentError(
-                        f"Endpoint '{endpoint_key}' is using the old APIHandlerConfig format. "
-                        f"The API has changed — each endpoint must include 'path' and 'http_method'. "
-                        f"Please update your stored config to the new format."
-                    )
-                body_mappings_dict = ep.get("input_body_mappings")
-                input_body_mappings = (
-                    BodyMappings.from_dict(body_mappings_dict)
-                    if body_mappings_dict
-                    else None
-                )
-                output_body_mappings_dict = ep.get("output_body_mappings")
-                output_body_mappings = (
-                    BodyMappings.from_dict(output_body_mappings_dict)
-                    if output_body_mappings_dict
-                    else None
-                )
-                self._endpoints[endpoint_key] = EndpointConfig(
-                    path=ep.get("path", ""),
-                    http_method=ep.get("http_method", HTTPMethod.POST),
-                    action=ep.get("action", schemas.APIHandlerAction.ALLOW),
-                    description=ep.get("description"),
-                    input_body_mappings=input_body_mappings,
-                    output_body_mappings=output_body_mappings,
-                )
-
-    @staticmethod
-    def _validate_http_method(http_method: HTTPMethod | str) -> HTTPMethod:
-        """Validate and normalize the provided HTTP method.
-
-        :param http_method: HTTP method to validate (HTTPMethod enum or string)
-        :return: Normalized HTTPMethod enum value
-        :raises mlrun.errors.MLRunInvalidArgumentError: If method is not a valid HTTPMethod or string
-        """
-        if isinstance(http_method, HTTPMethod):
-            return http_method
-        if isinstance(http_method, str):
-            try:
-                return HTTPMethod(http_method.upper())
-            except ValueError:
-                raise mlrun.errors.MLRunInvalidArgumentError(
-                    f"Invalid HTTP method string '{http_method}'. "
-                    f"Valid values are: {', '.join(m.value for m in HTTPMethod)}"
-                ) from None
-        # Not HTTPMethod or str - reject with helpful error
-        raise mlrun.errors.MLRunInvalidArgumentError(
-            f"http_method must be an HTTPMethod enum or string, got {type(http_method).__name__} "
-            f"with value '{http_method}'. Valid values are: {', '.join(m.value for m in HTTPMethod)}"
-        )
-
-    def get_endpoint_config(
-        self, method: HTTPMethod | str, path: str
-    ) -> "EndpointConfig | None":
-        """Get endpoint configuration for a specific method and path."""
-        method = self._validate_http_method(method)
-        path = EndpointConfig._normalize_path(path)
-        endpoint_key = serving_utils.combine_serving_endpoint_key(method, path)
-        return self._endpoints.get(endpoint_key)
-
-    def add_endpoint_handler(
-        self,
-        path: str,
-        http_method: HTTPMethod | str = HTTPMethod.POST,
-        action: schemas.APIHandlerAction = schemas.APIHandlerAction.ALLOW,
-        description: str | None = None,
-        input_body_mappings: "BodyMappings | None" = None,
-        output_body_mappings: "BodyMappings | None" = None,
-    ) -> None:
-        """Add an endpoint handler configuration.
-
-        :param path: URL path for the endpoint (e.g., ``/v1/models`` or ``/api/v1/*``)
-        :param http_method: HTTP method for the endpoint (``HTTPMethod`` enum or string like ``"GET"``, ``"POST"``)
-        :param action: Action to take for this endpoint (:py:class:`~mlrun.common.schemas.APIHandlerAction`)
-        :param description: Optional description of the endpoint
-        :param input_body_mappings: Optional input :class:`BodyMappings` for this endpoint (REST → graph).
-            If ``None``, the request body is passed through as-is.
-        :param output_body_mappings: Optional output :class:`BodyMappings` for this endpoint (graph → REST).
-            If ``None``, the response is returned as-is.
-        :raises mlrun.errors.MLRunValueError: If the path contains an invalid wildcard ``*`` pattern
-        """
-        ep = EndpointConfig(
-            path=path,
-            http_method=http_method,
-            action=action,
-            description=description,
-            input_body_mappings=input_body_mappings,
-            output_body_mappings=output_body_mappings,
-        )
-        endpoint_key = ep.get_endpoint_key()
-
-        # Warn if overriding an existing endpoint
-        if endpoint_key in self._endpoints:
-            logger.warning(
-                "Overriding existing endpoint handler configuration",
-                method=ep.http_method.value,
-                path=ep.path,
-                old_action=self._endpoints[endpoint_key].action,
-                new_action=str(action),
-            )
-
-        self._endpoints[endpoint_key] = ep
-
-    def remove_endpoint_handler(
-        self,
-        path: str,
-        http_method: HTTPMethod | str = HTTPMethod.POST,
-    ) -> None:
-        """Remove an endpoint handler configuration.
-
-        :param path: URL path for the endpoint to remove
-        :param http_method: HTTP method for the endpoint to remove (`HTTPMethod` enum or string like
-                            ``'GET'``, ``'POST'``)
-        """
-        http_method = self._validate_http_method(http_method)
-        path = EndpointConfig._normalize_path(path)
-        endpoint_key = serving_utils.combine_serving_endpoint_key(http_method, path)
-        self._endpoints.pop(endpoint_key, None)
-
-    def to_dict(self, fields=None, exclude=None, strip=False):
-        d = super().to_dict(fields=fields, exclude=exclude, strip=strip)
-        if d.get("endpoints"):
-            d["endpoints"] = {
-                k: v.to_dict(strip=strip) for k, v in self._endpoints.items()
-            }
-        return d
 
 
 def new_v2_model_server(
@@ -1488,7 +1107,7 @@ class ServingRuntime(nuclio_function.RemoteRuntime):
         Example::
 
             # Using APIHandlerConfig object
-            from mlrun.runtimes.nuclio.serving import APIHandlerConfig
+            from mlrun.serving.endpoint_mapping import APIHandlerConfig
             from mlrun.common.schemas.serving import APIHandlerAction
             from http import HTTPMethod
